GITHUB-APD-165: generate updates in a deterministic way
